// src/app/core/services/swagger-gen/profile/model/doctorOneDTO.ts
/**
 * Profile.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface DoctorOneDTO { 
    accountId?: string;
    firstName?: string;
    lastName?: string;
    middleName?: string;
    phoneNumber?: string;
    specializationTitle?: string;
    officeAddress?: string;
    statusTitle?: string;
    documentAbsolutUrl?: string;
    careerStartYear?: Date;
    birthDay?: Date;
}