// src/app/core/services/swagger-gen/profile/model/patientOneDTO.ts
/**
 * Profile.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface PatientOneDTO {
  accountId?: string;
  firstName?: string;
  lastName?: string;
  middleName?: string;
  phoneNumber?: string;
  birthDay?: Date;
  documentAbsolutUrl?: string;
  results: IResultForPatient[]
}

export interface IResultForPatient {
  title: string;
  children: IServiceForPatientResultDto[]
}

export interface IServiceForPatientResultDto {
  title: string;
  children: IAppointmentForPatientDto[]
}

export interface IAppointmentForPatientDto {
  date: string;
  doctorFullName: string;
  children: IResultDto[]
}

export interface IResultDto {
  absoluteUrl: string;
  fileTitle: string;
  children?: []
}