// src/app/core/services/swagger-gen/appointment/model/createAppointmentResultDto.ts
/**
 * Appoitment.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface CreateAppointmentResultDto { 
    appointmentId?: number;
    patientId?: number;
    patientFullName?: string;
    serviceTitle?: string;
    doctorFullName?: string;
    specializationTitle?: string;
    complaints?: string;
    conclusion?: string;
    recomendation?: string;
}